requirements.txt
pandas
openai
chromadb
gradio
python-dotenv